+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
  - Fix linking of some function return annotations.
  - Refine rendering of function signatures. Syntax errors are now handled more gracefully.
+ - Gracefully handle the case when users specify objects instead of strings in `__all__`.
 
 # 2022-05-15: pdoc 12.0.0